Updated example test cases to show error in test application
Added a test case for the Keyboard Down and Up Events labels in the test application to show an error in it.
Extended existing keywords "Press key" and "Press keys" to include press and release events to enable the test.

In the test application "WpfApplication.exe" in the "Keyboard Controls" tab, the Keyboard Down Events label does not work for some keys. E.g. Arrow Keys
(The Keyboard Up Events work)